requirements.txt
# Core
python-telegram-bot==21.5
telethon==1.36.0
apscheduler==3.10.4

# Fetchers
feedparser==6.0.11
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
trafilatura==1.12.2

# LLM / AI
openai==1.52.0

# Database
aiosqlite==0.20.0

# Utils
pyyaml==6.0.2
python-dotenv==1.0.1
Pillow==10.4.0
aiofiles==24.1.0
langdetect==1.0.9
